PR - Added changes for fetching amount and proc execution

+++ RecoPortal/src/main/java/com/lichfl/dao/RecoMatchRepo.java
@@ -13,25 +13,25 @@ public class RecoMatchRepo {
 	@PersistenceContext
 	private EntityManager entityManager;
 
-	public void executeMatchProc(int broKeyNo, int matchKey, String narration, int amount, String matchType) {
-	        StoredProcedureQuery storedProcedure = entityManager.createStoredProcedureQuery("hfl_brs.BRS_MATCH_BANK_RECO");
+	public void executeMatchProc(int broKeyNo, int matchKey, String narration, double amount, String matchType) {
+		StoredProcedureQuery storedProcedure = entityManager.createStoredProcedureQuery("hfl_brs.BRS_MATCH_BANK_RECO");
 
-	        // Register input parameters
-	        storedProcedure.registerStoredProcedureParameter("broKeyNo", Integer.class, ParameterMode.IN);
-	        storedProcedure.registerStoredProcedureParameter("matchKey", Integer.class, ParameterMode.IN);
-	        storedProcedure.registerStoredProcedureParameter("narration", String.class, ParameterMode.IN);
-	        storedProcedure.registerStoredProcedureParameter("amount", Integer.class, ParameterMode.IN);
-	        storedProcedure.registerStoredProcedureParameter("matchType", String.class, ParameterMode.IN);
+		// Register input parameters
+		storedProcedure.registerStoredProcedureParameter("broKeyNo", Integer.class, ParameterMode.IN);
+		storedProcedure.registerStoredProcedureParameter("matchKey", Integer.class, ParameterMode.IN);
+		storedProcedure.registerStoredProcedureParameter("narration", String.class, ParameterMode.IN);
+		storedProcedure.registerStoredProcedureParameter("amount", Double.class, ParameterMode.IN);
+		storedProcedure.registerStoredProcedureParameter("matchType", String.class, ParameterMode.IN);
 
-	        // Set parameter values
-	        storedProcedure.setParameter("broKeyNo", broKeyNo);
-	        storedProcedure.setParameter("matchKey", matchKey);
-	        storedProcedure.setParameter("narration", narration);
-	        storedProcedure.setParameter("amount", amount);
-	        storedProcedure.setParameter("matchType", matchType);
+		// Set parameter values
+		storedProcedure.setParameter("broKeyNo", broKeyNo);
+		storedProcedure.setParameter("matchKey", matchKey);
+		storedProcedure.setParameter("narration", narration);
+		storedProcedure.setParameter("amount", amount);
+		storedProcedure.setParameter("matchType", matchType);
 
-	        // Execute the stored procedure
-	
-	        storedProcedure.execute();
+		// Execute the stored procedure
+
+		storedProcedure.execute();
 	}
 }
